+++ src/util/topology.h
@@ -15,19 +15,19 @@ namespace topo{
 
 class rect{
 public:
-  rect(MPI_Comm comm, size_t c, size_t num_chunks=0){
+  rect(MPI_Comm comm, int64_t c, int64_t num_chunks=0){
 
     this->num_chunks = num_chunks;
     MPI_Comm column;
     int rank, size, columnRank, cubeRank;
     MPI_Comm_rank(comm, &rank);
     MPI_Comm_size(comm, &size);
-    size_t SubCubeSize = c*c*c;
-    size_t SubCubeSliceSize = c*c;
+    int64_t SubCubeSize = c*c*c;
+    int64_t SubCubeSliceSize = c*c;
     MPI_Comm_split(comm, rank/SubCubeSize, rank, &this->cube);
     MPI_Comm_rank(this->cube, &cubeRank);
-    size_t temp1 = (cubeRank%c) + c*(cubeRank/SubCubeSliceSize);
-    size_t temp2 = rank % SubCubeSliceSize;
+    int64_t temp1 = (cubeRank%c) + c*(cubeRank/SubCubeSliceSize);
+    int64_t temp2 = rank % SubCubeSliceSize;
     MPI_Comm_split(this->cube, cubeRank/c, cubeRank, &this->depth);
     MPI_Comm_split(this->cube, temp1, cubeRank, &this->row);
     // Note: columnComm in the tunable grid is of size d, not size c like in the 3D grid
@@ -59,12 +59,12 @@ public:
   }
 
   MPI_Comm world,row,column_contig,column_alt,depth,slice,cube;
-  size_t c,d,x,y,z,num_chunks;
+  int64_t c,d,x,y,z,num_chunks;
 };
 
 class square{
 public:
-  square(MPI_Comm comm, size_t c, size_t num_chunks=0){
+  square(MPI_Comm comm, int64_t c, int64_t num_chunks=0){
 
     this->num_chunks = num_chunks;
     int rank,size;
@@ -73,7 +73,7 @@ public:
 
     this->c = c;
     this->d = std::nearbyint(std::ceil(pow(size/c,1./2.)));
-    size_t TopFaceSize = this->d*this->d;
+    int64_t TopFaceSize = this->d*this->d;
     this->z = rank%c;
     this->y = rank/TopFaceSize;
     this->x = (rank%TopFaceSize)/this->c;
@@ -101,7 +101,7 @@ public:
   }
 
   MPI_Comm world,row,column,slice,depth;
-  size_t c,d,x,y,z,num_chunks;
+  int64_t c,d,x,y,z,num_chunks;
 };
 }
